move json parsing to base Ministry object
