+++ app/src/main/java/com/expidev/gcmapp/http/GmaApiClient.java
@@ -220,7 +220,7 @@ public final class GmaApiClient extends AbstractTheKeyApi<AbstractTheKeyApi.Requ
 
             // is this a successful response?
             if (conn.getResponseCode() == HttpURLConnection.HTTP_OK) {
-                return AssociatedMinistry.listFromJson(new JSONArray(IOUtils.readString(conn.getInputStream())));
+                return Ministry.listFromJson(new JSONArray(IOUtils.readString(conn.getInputStream())));
             }
         } catch (final JSONException e) {
             Log.e(TAG, "error parsing getAllMinistries response", e);
